Extract file copy function & refactor init function
